Rename resource to `verticalpodautoscaler`
It could potentially be used to create multiple VPA resources, not only
for Prometheus.
