+++ service/controller/resource/verticalpodautoscaler/test/case-0-cluster-api.golden
