+++ service/controller/resource/verticalpodautoscaler/test/case-1-awsconfig.golden
