+++ service/controller/resource/verticalpodautoscaler/test/case-2-azureconfig.golden
