+++ service/controller/resource/verticalpodautoscaler/test/case-3-kvmconfig.golden
