+++ service/controller/resource/verticalpodautoscaler/test/case-4-control-plane.golden
