+++ service/controller/resource/verticalpodautoscaler/test/case-5-cluster-api-v1alpha3.golden
